+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.32.0...master)
-
+### Changed
+- Moved graph-server to it's own repo <https://github.com/usgs/wdfn-graph-server>
 
 ## [0.32.0](https://github.com/usgs/waterdataui/compare/waterdataui-0.32.0...master)
 ### Added